Add gender filter to the teacher search function ,Modify scoring data type

+++ evaluate/models.py
@@ -8,4 +8,4 @@ class Table(models.Model):
     evaluated_usr_email = models.CharField(max_length=100)  #被评价人邮箱
     evaluate_usr_name = models.CharField(max_length=100)  #评价人昵称
     evaluate_content = models.CharField(max_length=300) #评价内容
-    evaluate_score = models.IntegerField(default=5) #评分(1,2,3,4,5 分)
+    evaluate_score = models.FloatField(default=5) #评分(1,2,3,4,5 分)
